+++ spotify_wordcloud/api/image.py
@@ -1,66 +1,47 @@
+import boto3
 from flask import *
 from flask_dance.contrib.spotify import spotify
-from wordcloud import WordCloud
 import tweepy
-import boto3
+from wordcloud import WordCloud
 
-from os import path
 from datetime import date, datetime
-import logging
 import hashlib
+import logging
+from os import path
 
-from spotify_wordcloud import app, db
-from .models import Pictures
+from spotify_wordcloud import db
+from spotify_wordcloud.models import Pictures
 
 s3_client = boto3.client('s3', region_name='ap-northeast-1')
 
 
-
-@app.route("/")
-def index():
-    if spotify.authorized:
-        if 'oauth_verifier' in session:
-            return render_template('authorized.html', twitter_authorized=True)
-        else:
-            return render_template('authorized.html', twitter_authorized=False)
-    else:
-        return render_template('unauthorized.html')
+app = Blueprint('image', __name__, url_prefix='/')
 
 
-@app.route("/login")
-def login():
-    return redirect(url_for("spotify.login"))
+def hash_generation(session, text):
+    data = spotify.get("/v1/me/top/artists?limit=50").json()
+
+    for i, item in enumerate(data['items']):
+        for k in range((50-i)//10):
+            text += item['name']
+            text += " "
+
+    session["spotify_wordcloud_text"] = text  # save text
+    ha = hashlib.md5(text.encode('utf-8')).hexdigest()
+    session["spotify_wordcloud_hash"] = ha  # save hash
+
+    return text
 
 
-@app.route("/logout")
-def logout():
-    session.clear()
-    return redirect("/")
-
-
-@app.route('/twitter_auth')
-def twitter_auth():
-    redirect_url = ""
-    auth = tweepy.OAuthHandler(app.config["TWITTER_API_KEY"], app.config["TWITTER_API_SECRET"], 'https://spotify-wordcloud.herokuapp.com/callback')
-
-    try:
-        redirect_url = auth.get_authorization_url()
-        session['request_token'] = auth.request_token
-    except tweepy.TweepError as e:
-        logging.error(str(e))
-
-    return redirect(redirect_url)
-
-
-@app.route("/callback")
-def callback():
-    try:
-        session['oauth_verifier'] = request.args.get('oauth_verifier')
-        return redirect("/")
-
-    except Exception as e:
-        logging.error(str(e))
-        return Response(status=500)
+def image_generation(text, ha):
+    freq = WordCloud(stopwords=set()).process_text(text)
+    # force regularize
+    for k, v in freq.items():
+        if v >= 6:
+            freq[k] = 6
+    wc = WordCloud(font_path='/app/.fonts/ipaexg.ttf', width=1024, height=576, colormap='cool', stopwords=set()).generate_from_frequencies(freq)
+    image = wc.to_image()
+    image.save(f"/tmp/{ha}.png", format='png', optimize=True)
 
 
 @app.route('/generate')
@@ -68,18 +49,9 @@ def generate():
     if spotify.authorized:
         try:
             text = ""
-
+            
             if "spotify_wordcloud_text" not in session:
-                data = spotify.get("/v1/me/top/artists?limit=50").json()
-
-                for i, item in enumerate(data['items']):
-                    for k in range((50-i)//10):
-                        text += item['name']
-                        text += " "
-
-                session["spotify_wordcloud_text"] = text  # save text
-                ha = hashlib.md5(text.encode('utf-8')).hexdigest()
-                session["spotify_wordcloud_hash"] = ha  # save hash
+                text = hash_generation(session, text)
 
             text = session["spotify_wordcloud_text"]
             ha = session["spotify_wordcloud_hash"]
@@ -87,14 +59,7 @@ def generate():
             if path.exists(f"/tmp/{ha}.png"):
                 return send_file(f"/tmp/{ha}.png", mimetype='image/png')
 
-            freq = WordCloud(stopwords=set()).process_text(text)
-            # force regularize
-            for k, v in freq.items():
-                if v >= 6:
-                    freq[k] = 6
-            wc = WordCloud(font_path='/app/.fonts/ipaexg.ttf', width=1024, height=576, colormap='cool', stopwords=set()).generate_from_frequencies(freq)
-            image = wc.to_image()
-            image.save(f"/tmp/{ha}.png", format='png', optimize=True)
+            image_generation(text, ha)
 
             return send_file(f"/tmp/{ha}.png", mimetype='image/png')
 
@@ -113,28 +78,12 @@ def regenerate():
             text = ""
 
             if "spotify_wordcloud_text" not in session:
-                data = spotify.get("/v1/me/top/artists?limit=50").json()
-
-                for i, item in enumerate(data['items']):
-                    for k in range((50-i)//10):
-                        text += item['name']
-                        text += " "
-
-                session["spotify_wordcloud_text"] = text  # save text
-                ha = hashlib.md5(text.encode('utf-8')).hexdigest()
-                session["spotify_wordcloud_hash"] = ha  # save hash
+                text = hash_generation(session, text)
 
             text = session["spotify_wordcloud_text"]
             ha = session["spotify_wordcloud_hash"]
 
-            freq = WordCloud(stopwords=set()).process_text(text)
-            # force regularize
-            for k, v in freq.items():
-                if v >= 6:
-                    freq[k] = 6
-            wc = WordCloud(font_path='/app/.fonts/ipaexg.ttf', width=1024, height=576, colormap='cool', stopwords=set()).generate_from_frequencies(freq)
-            image = wc.to_image()
-            image.save(f"/tmp/{ha}.png", format='png', optimize=True)
+            image_generation(text, ha)
 
             return send_file(f"/tmp/{ha}.png", mimetype='image/png')
 
@@ -223,7 +172,7 @@ def tweet():
                 image.save(f"/tmp/{ha}.png", format='png', optimize=True)
 
             # Upload Media to Twitter
-            auth = tweepy.OAuthHandler(app.config["TWITTER_API_KEY"], app.config["TWITTER_API_SECRET"], 'https://spotify-wordcloud.herokuapp.com/callback')
+            auth = tweepy.OAuthHandler(current_app.config["TWITTER_API_KEY"], current_app.config["TWITTER_API_SECRET"], 'https://spotify-wordcloud.herokuapp.com/callback')
             token = session.pop('request_token', None)
             auth.request_token = token
             verifier = session.pop('oauth_verifier', None)
@@ -240,40 +189,5 @@ def tweet():
         except Exception as e:
             logging.error(str(e))
             return render_template('result.html', result="ツイートに失敗しました。")
-    else:
-        return Response(status=401)
-
-
-@app.route('/history')
-def history():
-    if spotify.authorized:
-        try:
-            pictures = db.session.query(Pictures).filter(Pictures.user_id==session["user_id"]).order_by(Pictures.created_at.desc()).all()
-            db.session.commit()
-
-            return render_template('history.html', pictures=pictures)
-
-        except Exception as e:
-            logging.error(str(e))
-            return render_template('result.html', result="画像一覧の表示に失敗しました。時間を置いて再度試してみてください。")
-
-    else:
-        return Response(status=401)
-
-
-@app.route('/history/<string:file_hash>', methods=['POST'])
-def delete_picture(file_hash):
-    if spotify.authorized and (request.form.get('_method') == 'DELETE'):
-        try:
-            pictures = db.session.query(Pictures).filter(Pictures.file_name==(file_hash+".png"))\
-                .filter(Pictures.user_id==session["user_id"]).delete()
-            db.session.commit()
-
-            return redirect("/history")
-
-        except Exception as e:
-            logging.error(str(e))
-            return render_template('result.html', result="画像の削除に失敗しました。時間を置いて再度試してみてください。")
-
     else:
         return Response(status=401)
