+++ spotify_wordcloud/templates/history.html
@@ -4,24 +4,30 @@
 <h2>過去に作成した画像一覧</h2>
 <br/>
 <div class="row">
-{% for item in pictures %}
-  {% set imgsrc = "https://spotify-wordcloud.s3-ap-northeast-1.amazonaws.com/" + item.file_name %}
-  {% set deletesrc = "/history/" + item.file_name.replace('.png', '') %}
+{% if pictures|length > 0 %}
+  {% for item in pictures %}
+    {% set imgsrc = "https://spotify-wordcloud.s3-ap-northeast-1.amazonaws.com/" + item.file_name %}
+    {% set deletesrc = "/history/" + item.file_name.replace('.png', '') %}
 
-  <div class="col-md-6">
-    <div class="card mb-4 shadow-sm">
-      <img src={{imgsrc}} width="100%">
-      <div class="card-body">
-        <p class="card-text">{{ item.created_at.strftime('%Y年%m月%d日 %H:%M:%S') }} 作成</p>
-        <form action={{deletesrc}} method="POST">
-          <input type="hidden" name="_method" value="DELETE"/>
-          <input type="hidden" name="_csrf_token" value="{{ csrf_token() }}">
-          <button type="submit" class="btn btn-default" style="text-align: right; color: red; margin: 0; padding: 0;">削除</button>
-        </form>
+    <div class="col-md-6">
+      <div class="card mb-4 shadow-sm">
+        <img src={{imgsrc}} width="100%">
+        <div class="card-body">
+          <p class="card-text">{{ item.created_at.strftime('%Y年%m月%d日 %H:%M:%S') }} 作成</p>
+          <form action={{deletesrc}} method="POST">
+            <input type="hidden" name="_method" value="DELETE"/>
+            <input type="hidden" name="_csrf_token" value="{{ csrf_token() }}">
+            <button type="submit" class="btn btn-default" style="text-align: right; color: red; margin: 0; padding: 0;">削除</button>
+          </form>
+        </div>
       </div>
     </div>
+  {% endfor %}
+{% else %}
+  <div class="col-md-6">
+    <p>まだ画像が保存されていません。</p>
   </div>
-{% endfor %}
+{% endif %}
 </div>
 <div class="border border-light p-3 mb-4">
